Splits lib and cli
