+++ WordscapesCheat.sln
@@ -3,7 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
 VisualStudioVersion = 15.0.28307.572
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WordscapesCheat", "src\WordscapesCheat.csproj", "{FF29FD59-7363-4B26-A24B-92FDDD8A2E51}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WordscapesCheatCLI", "src\WordscapesCheat.CLI\WordscapesCheat.CLI.csproj", "{FF29FD59-7363-4B26-A24B-92FDDD8A2E51}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WordscapesCheat", "src\WordscapesCheat\WordscapesCheat.csproj", "{A3415268-B8C1-4A07-AF24-EF2CFD2C7E95}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -15,6 +17,10 @@ Global
 		{FF29FD59-7363-4B26-A24B-92FDDD8A2E51}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{FF29FD59-7363-4B26-A24B-92FDDD8A2E51}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{FF29FD59-7363-4B26-A24B-92FDDD8A2E51}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A3415268-B8C1-4A07-AF24-EF2CFD2C7E95}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A3415268-B8C1-4A07-AF24-EF2CFD2C7E95}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A3415268-B8C1-4A07-AF24-EF2CFD2C7E95}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A3415268-B8C1-4A07-AF24-EF2CFD2C7E95}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
