+++ src/WordscapesCheat.CLI/App.config
