+++ src/WordscapesCheat.CLI/Properties/AssemblyInfo.cs
@@ -5,12 +5,12 @@ using System.Runtime.InteropServices;
 // General Information about an assembly is controlled through the following
 // set of attributes. Change these attribute values to modify the information
 // associated with an assembly.
-[assembly: AssemblyTitle("WordscapesCheatCopy")]
+[assembly: AssemblyTitle("WordscapesCheat.CLI")]
 [assembly: AssemblyDescription("")]
 [assembly: AssemblyConfiguration("")]
 [assembly: AssemblyCompany("")]
-[assembly: AssemblyProduct("WordscapesCheatCopy")]
-[assembly: AssemblyCopyright("Copyright ©  2019")]
+[assembly: AssemblyProduct("WordscapesCheat.CLI")]
+[assembly: AssemblyCopyright("Copyright © KeenanCronyn 2019")]
 [assembly: AssemblyTrademark("")]
 [assembly: AssemblyCulture("")]
