+++ src/WordscapesCheat/CheatFunctions.cs
@@ -6,16 +6,8 @@ using System.Threading.Tasks;
 
 namespace WordscapesCheat
 {
-    class Program
-    {        
-        static void Main(string[] args)
-        {
-            string characterSet = "catmbotle";
-            List<string> fakeDictionary = new List<string>() { "cat", "mat", "cab", "catm" };
-            
-            PrintMatchingWords(BuildMatchingWordsArray(characterSet, fakeDictionary));
-        }
-
+    public static class CheatFunctions
+    {
         public static int[] BuildOccurenceArray(string inputString)
         {
             const int A_ASCII = 97;
@@ -37,14 +29,14 @@ namespace WordscapesCheat
             int[] givenLettersArray = BuildOccurenceArray(givenLetters);
 
             foreach (string word in dictionary)
-            { 
+            {
                 if (word.Length > 2 && word.Length < 8)
                 {
                     int[] wordArray = BuildOccurenceArray(word);
 
                     if (TestIfMatching(givenLettersArray, wordArray) == true)
                         matchingWordsList.Add(word);
-                }                                    
+                }
             }
 
             string[] matchingWordsArray = matchingWordsList.ToArray();
@@ -78,25 +70,9 @@ namespace WordscapesCheat
             for (int i = 0; i < givenLettersArray.Length; i++)
             {
                 if (wordArray[i] > givenLettersArray[i])
-                    return false;                                
+                    return false;
             }
             return true;
         }
-
-        public static void PrintMatchingWords(string[] matchingWordsArray)
-        {
-            // this prints out the words in the array if successful. 
-            // I could include something that prints if nothing is in the array.
-            if (matchingWordsArray.Length == 0)
-                Console.WriteLine("No words could be found that match what you asked for");
-            else
-            {
-                foreach (var word in matchingWordsArray)
-                {
-                    Console.WriteLine(word);
-                }
-            }            
-        }
-
     }
 }
